% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_distribution.R
\name{asm_distribution}
\alias{asm_distribution}
\title{Test the assumption of a normal distribution}
\usage{
asm_distribution(
  data = NULL,
  tests = NULL,
  report = c("pvalue", "statistic")[1],
  simulate = FALSE,
  checkData = !simulate,
  testBatch = NULL
)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}

\item{tests}{a selection of tests via its names or index. See also the \link[assumptions]{asm_library} function.}

\item{report}{which statistics should be reported.}

\item{simulate}{logical if running in simulation.}

\item{checkData}{logical if TRUE data is checked before running tests.}

\item{testBatch}{a subset of \link[assumptions]{asm_library} table for selected distribution tests.}
}
\description{
This function applies tests for the normal distribution assumption.
}
\examples{
# normal distribution
asm_distribution(rnorm(100))
# gamma distribution
asm_distribution(rgamma(100, 2))
}
