% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_graph.R
\name{asm_graphBoxplot}
\alias{asm_graphBoxplot}
\title{Plot data as Boxplot}
\usage{
asm_graphBoxplot(data, legend = T)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}

\item{legend}{logical add legend.}
}
\description{
This function plots data as boxplot.
}
\examples{
asm_graphBoxplot(iris[,c("Sepal.Width", "Sepal.Length")])
}
