% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_graph.R
\name{asm_graphCumvar}
\alias{asm_graphCumvar}
\title{Plot Cummulative Variance of data}
\usage{
asm_graphCumvar(data)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}
}
\description{
This function plots the fluctuation of the variance if its calculated as cummulated variance.
}
\examples{
# normal data vs. mixed non-id data
asm_graphCumvar(list(N = rnorm(60), M = c(rnorm(20), rnorm(20, sd = 2), rnorm(20, sd = 10))))
}
