% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_graph.R
\name{asm_graphDifferencesign}
\alias{asm_graphDifferencesign}
\title{Difference-Sign Plot of data}
\usage{
asm_graphDifferencesign(data)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}
}
\description{
This function plots the relative shares of runs for the differences. Every value is subtracted by
its predecessor and counted if above zero. The relative share of a change of directions (run) is
plotted. The Difference-Sign plot is a graphical representation of the Difference-Sign Runs Test.
See also the \link[randtests]{difference.sign.test} function.
}
\examples{
asm_graphDifferencesign(list(A = rnorm(100), B = (1:100/10)+rnorm(10)))
}
