% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_graph.R
\name{asm_graphDotplot}
\alias{asm_graphDotplot}
\title{Plot data as dotplot}
\usage{
asm_graphDotplot(data)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}
}
\description{
This function plots data as dotplot.
}
\examples{
asm_graphDotplot(iris[,c("Sepal.Width", "Sepal.Length")])
}
