% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_graph.R
\name{asm_graphEcdf}
\alias{asm_graphEcdf}
\title{Plot Cummulative Distribution Function of data}
\usage{
asm_graphEcdf(data, theoMean = mean(data), theoSD = sd(data))
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}

\item{theoMean}{mean for theoretic distribution.}

\item{theoSD}{standard deviation for theoretic distribution.}
}
\description{
This function plots the empiricial cummulative distribution function derived from data against
the theoretical cummulative distribution function of the normal distribution with mean and sd
calculated from data.
}
\examples{
asm_graphEcdf(iris[,c("Sepal.Width", "Sepal.Length")])
}
