% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_graph.R
\name{asm_graphHist}
\alias{asm_graphHist}
\title{Plot Histogramm of data}
\usage{
asm_graphHist(data, addLine = TRUE)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}

\item{addLine}{Draw curve of normal distribution with mean and sd calculated from data.}
}
\description{
This function plots data as histogramm.
}
\examples{
asm_graphHist(iris[,c("Sepal.Width", "Sepal.Length")])
}
