% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_graph.R
\name{asm_graphPPlot}
\alias{asm_graphPPlot}
\title{Probability–Probability Plot of data}
\usage{
asm_graphPPlot(data, theoMean = mean(data), theoSD = sd(data))
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}

\item{theoMean}{mean for theoretic distribution.}

\item{theoSD}{standard deviation for theoretic distribution.}
}
\description{
This function plots the Probability–Probability plot of data.
}
\examples{
asm_graphPPlot(iris[,c("Sepal.Width", "Sepal.Length")])
}
