% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_graph.R
\name{asm_graphQQplot}
\alias{asm_graphQQplot}
\title{Quantile-Quantile Plot of data}
\usage{
asm_graphQQplot(data)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}
}
\description{
This function plots the Quantile-Quantile plot of data.
}
\examples{
asm_graphQQplot(iris[,c("Sepal.Width", "Sepal.Length")])
}
