% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_graph.R
\name{asm_graphScatterplot}
\alias{asm_graphScatterplot}
\title{Scatterplot of data}
\usage{
asm_graphScatterplot(data)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}
}
\description{
This function plots the scatterplot of data.
}
\examples{
normal data vs. mixed non-id data
asm_graphScatterplot(list(N = rnorm(60), M = c(rnorm(20), rnorm(20, 2), rnorm(20, 10))))
}
