% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_graph.R
\name{asm_graphWindow}
\alias{asm_graphWindow}
\title{Window Graph of data}
\usage{
asm_graphWindow(data)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}
}
\description{
This function draws 'windows' onto a scatterplot of the data as a support to imagen homogeneity.
Each window should contain the same pattern if the data points are homogeneous or
identically distributed.
}
\examples{
# normal data vs. mixed non-id data
asm_graphWindow(list(N = rnorm(60), M = c(rnorm(20), rbeta(20, 2, 5), rt(20, 9))))
}
