% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_homogeneity.R
\name{asm_homogeneity}
\alias{asm_homogeneity}
\title{Test the assumption of identical distribution}
\usage{
asm_homogeneity(
  data = NULL,
  tests = NULL,
  report = c("pvalue"),
  testBatch = NULL,
  simulate = FALSE,
  checkData = !simulate
)
}
\arguments{
\item{data}{a data.frame with at least two columns or a list of length two.}

\item{tests}{a selection of tests via its names or index. See also the \link[assumptions]{asm_library} function.}

\item{report}{which statistics should be reported.}

\item{testBatch}{a subset of \link[assumptions]{asm_library} table for selected homogeneity tests.}

\item{simulate}{logical if running in simulation.}

\item{checkData}{logical if TRUE data is checked before running tests.}
}
\description{
This function applies tests for identically distribution assumptions of two variables.
}
\examples{
asm_homogeneity(data.frame(X=rnorm(50), Y=rnorm(50, mean = 2, sd = 2)))
}
