% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_independence.R
\name{asm_independence}
\alias{asm_independence}
\title{Test the assumption of independence}
\usage{
asm_independence(
  data = NULL,
  tests = NULL,
  report = c("p.value", "statistic", "parameter")[1],
  testBatch = NULL,
  simulate = FALSE,
  lag = 1,
  checkData = !simulate
)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}

\item{tests}{a selection of tests via its names or index. See also the \link[assumptions]{asm_library} function.}

\item{report}{which statistics should be reported.}

\item{testBatch}{a subset of \link[assumptions]{asm_library} table for selected independence tests.}

\item{simulate}{logical if running in simulation.}

\item{lag}{logical if running in simulation.}

\item{checkData}{logical if TRUE data is checked before running tests.}
}
\description{
This function applies tests for independece assumption in form of autocorrelation.
}
\examples{
asm_independence(list(A = 1:100, B = 1:100))
}
