% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_internals.R
\name{asm_intCheckData}
\alias{asm_intCheckData}
\title{Check input data for asm_graph functions}
\usage{
asm_intCheckData(data, return = T)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}

\item{return}{logical. Return data.}
}
\description{
This function tests a vector if it containts numeric data with at least two not NA datapoints.
}
\examples{
asm_intCheckData()
}
\keyword{internal}
