% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_internals.R
\name{asm_internalLibNames}
\alias{asm_internalLibNames}
\title{Appends pretty names to asm_library()}
\usage{
asm_internalLibNames(libTable)
}
\arguments{
\item{libTable}{integer samplesize to draw.}
}
\description{
This function appends full names and links to their documentation of tests in asm_library() table.
}
\examples{
asm_library(print = T)
}
\keyword{internal}
