% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_internals.R
\name{asm_intHomogTest}
\alias{asm_intHomogTest}
\title{Apply homogeneity test in asm_homogeneity}
\usage{
asm_intHomogTest(
  data,
  test = c("ftest", "bartlett", "levene", "oneillmathews")[1]
)
}
\arguments{
\item{data}{a data.frame with at least two columns or a list of length two.}

\item{tests}{vector of tests.}
}
\description{
This function is used in asm_homogeneity to conduct homogeneity tests.
}
\examples{
asm_intHomogTest(list(A = rnorm(30), B = rnorm(40)), "levene")
}
\keyword{internal}
