% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_internals.R
\name{asm_intListToTab}
\alias{asm_intListToTab}
\title{Convert list to data.table}
\usage{
asm_intListToTab(li, suffix = LETTERS[1:length(li)], dim = length(li))
}
\arguments{
\item{li}{integer samplesize to draw.}

\item{suffix}{distribution function to draw sample with.}
}
\description{
This function converts a list of arguments for sim_func of asm_simulate to a data.table.
}
\examples{
asm_intListToTab(list(list(mean = 1, sd = 1)))
}
\keyword{internal}
