% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_internals.R
\name{asm_intSelT}
\alias{asm_intSelT}
\title{Select pre-test}
\usage{
asm_intSelT(testTable, testSelection)
}
\arguments{
\item{testTable}{asm_library() table.}

\item{testSelection}{name given in 'Name' column or index of asm_library() table.}
}
\description{
This function selects a pre-test from asm_library() table for usage in asm_ functions.
}
\examples{
asm_intSelT(asm_library(), "Runs")
asm_intSelT(asm_library(), c(3, 33))
}
\keyword{internal}
