% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_library.R
\name{asm_library}
\alias{asm_library}
\title{List of all assumptions tests}
\usage{
asm_library(index = NULL, print = TRUE)
}
\arguments{
\item{index}{a integer indicating the row number of data.table to open documentation for indexed function.}

\item{print}{logical. If TRUE print also the Name, Package, Link to documentation and the name of the
reference column for \link[assumptions]{asm_simulate} results.}
}
\description{
This function returns a data.table with all pre-tests and graphs to assess statistical assumptions.
}
\examples{
asm_library()
# shortcut to open documantation for 33rd test
asm_library(33)
}
