% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_pre.R
\name{asm_preJoint}
\alias{asm_preJoint}
\title{Spanos Auxiliary Regression}
\usage{
asm_preJoint(data, simulate = FALSE)
}
\arguments{
\item{data}{a data.frame with at least two columns or a list of length two.}

\item{simulate}{logical if running in simulation.}
}
\description{
This function conducts one version of Aris Spanos Auxiliary Regression to test the iid assumption of a given
variable.
}
\examples{
# test with normal data
asm_preJoint(rnorm(40))
# test with non identically distributed data
asm_preJoint(c(rnorm(20), rnorm(20, 2, 2)))
# test with non independent data
asm_preJoint(diffinv(rnorm(40)))
}
\references{
Spanos, A. (2019). \emph{Misspecification (M-S) Testing}. Chapter 15 of \emph{Probability theory and statistical inference: Empirical modeling with observational data.}
}
