% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_pre.R
\name{asm_preTests}
\alias{asm_preTests}
\title{Apply assumption tests}
\usage{
asm_preTests(data, tests = NULL, testBatch = NULL, simulate = FALSE)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}

\item{tests}{a selection of tests via its names or index.}

\item{testBatch}{a subset of \link[assumptions]{asm_library} table for selected tests.}

\item{simulate}{logical if running in simulation.}
}
\description{
This function applies all tests of \link[assumptions]{asm_library} tests or a defined selection of tests.
}
\examples{
# test with normal data
asm_preTests(rnorm(40))
# test with non identically distributed data
asm_preTests(c(rnorm(20), rnorm(20, 2, 2)))
# test with non independent data
asm_preTests(diffinv(rnorm(40)))
# with homogeneity tests
asm_preTests(list(A = rnorm(20), B = rnorm(20, 2, 2)))
}
