% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_randomness.R
\name{asm_randomness}
\alias{asm_randomness}
\title{Test the assumption of randomness}
\usage{
asm_randomness(
  data = NULL,
  tests = NULL,
  report = c("p.value", "statistic")[1],
  testBatch = NULL,
  simulate = FALSE,
  checkData = !simulate
)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}

\item{tests}{a selection of tests via its names or index. See also the \link[assumptions]{asm_library} function.}

\item{report}{which statistics should be reported.}

\item{testBatch}{a subset of \link[assumptions]{asm_library} table for selected randomness tests.}

\item{simulate}{logical if running in simulation.}

\item{checkData}{logical if TRUE data is checked before running tests.}
}
\description{
This function applies tests for the randomness assumption.
}
\examples{
# non random brownian motion
asm_randomness(diffinv(rnorm(999)))
# random normal data
asm_randomness(rnorm(999))
}
