% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_report.R
\name{asm_reportData}
\alias{asm_reportData}
\title{Describe data in simulation}
\usage{
asm_reportData(data, with = c("mean", "median", "sd", "skew", "kurtosis"))
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}

\item{with}{which descriptive parameters to calculate.}
}
\description{
This function is used in simulation to calculate descriptive parameters of simulated data.
}
\examples{
asm_reportData(rnorm(10))
}
