% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_report.R
\name{asm_reportSim}
\alias{asm_reportSim}
\title{Describe simulation result}
\usage{
asm_reportSim(
  simResult,
  digits = 4,
  report = c("result", "cross", "description")
)
}
\arguments{
\item{simResult}{simulation result of \link[assumptions]{asm_simulate} function.}

\item{digits}{integer indicating the number of decimal places for rounding the results.}

\item{report}{what aggregations should be reported. "result" gives aggregation for all pre- and post-tests and
strategies. "cross" gives a cross tabulation for all tests. "description" gives aggregation for all descritpive parameters.}
}
\description{
This function can be used to aggregate a simulation result of the \link[assumptions]{asm_simulate} function.
}
\examples{
asm_reportSim(asm_simulate())
}
