% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_simulate.R
\name{asm_simData}
\alias{asm_simData}
\title{Simulate multiple variables}
\usage{
asm_simData(
  N = c(30, 40),
  distr = c("rnorm", "rnorm"),
  dots = rep(list(list()), length(distr)),
  suffix = c("X", "Y")
)
}
\arguments{
\item{N}{integer sample size to draw.}

\item{distr}{distribution function to draw sample with.}

\item{dots}{args passed to distribution function.}

\item{suffix}{how to name simulated variables.}
}
\description{
This function simulates multiple variables with given distribution functions.
}
\examples{
asm_simData()
}
