% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_simulate.R
\name{asm_simPostTest}
\alias{asm_simPostTest}
\title{Conduct hypothesis test in simulation}
\usage{
asm_simPostTest(
  data,
  paired = FALSE,
  report = c("p.value", "conf.int", "statistic")[1]
)
}
\arguments{
\item{data}{a vector, list or data.frame of data values.}

\item{paired}{logical - conduct a paired hypothesis test.}

\item{report}{which statistics should be saved in simulation.}
}
\description{
This function is used in asm_simulate() to calculate hypothesis test like t-test, welch or wilcoxon.
}
\examples{
asm_simPostTest(rnorm(10))
asm_simPostTest(iris[, c("Sepal.Width", "Sepal.Length")])
}
