% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_simulate.R
\name{asm_simulate}
\alias{asm_simulate}
\title{Simulate data for asm_ functions}
\usage{
asm_simulate(
  simulations = 1,
  sim_n = c(10, 30),
  sim_func = "asm_simData",
  sim_args = list(),
  describe = asm_reportData,
  pre_test = asm_preTests,
  pre_selection = NULL,
  post_test = asm_simPostTest,
  strategies = list(),
  showSimPlan = FALSE,
  alpha = 0.05,
  useParallel = FALSE,
  dryRun = TRUE
)
}
\arguments{
\item{simulations}{integer indicating the number of simulations per configuration.}

\item{sim_n}{integer or list of integers for samplesize(s) to draw.}

\item{sim_func}{character a function name as character to use for drawing sample.}

\item{sim_args}{list arguments for sim_func to define different simulations.}

\item{describe}{function to use for descriptive parameters of each simulated sample.}

\item{pre_test}{function to use for conducting pre-tests.}

\item{pre_selection}{character vector containing the 'Names' of pre-tests given by \link[assumptions]{asm_library} to select subset of pre-tests.}

\item{post_test}{function to conduct post-test with.}

\item{strategies}{list of quotes to define decision strategies for selecting the post-test depending on pre-test (if certain
assumptions are not met use a different post-test). Quotes must contain the 'SimulationName' of \link[assumptions]{asm_library}.}

\item{showSimPlan}{logical. Set TRUE to preview all configurations of simulation.}

\item{alpha}{numeric indicating the global alpha-level for pre- and post-tests.}

\item{useParallel}{logical. If TRUE running on multiple cores using \link[parallel]{mclapply}.}

\item{dryRun}{logical. If TRUE test a dry run by simulating each configuration ones before running all simulations.}
}
\value{
Return value is a data.table containing the results for each simulation. Each simulation is corresponding to a row
in the result table.
\describe{
     \item{'sim_'}{All columns beginning with sim_ prefix are describing the configuration of the simulation.}
     \item{'descr_'}{All columns beginning with descr_ prefix are the result of the describe function of the drawn sample (default is \link[assumptions]{asm_reportData}).}
     \item{'pre_'}{All columns beginning with 'pre_' prefix are containing the results of the selected pre-tests with the selected report statistics (default is the p-value). Columns with 'SimulationName' of \link[assumptions]{asm_library} are a logical value indicating if p-value of pre-test is smaller than global alpha.}
     \item{'post_'}{All columns beginning with post_ prefix arecontaining the results of the selected post-tests with the selected report statistics (default is the p-value). Columns with only the name of the test (like post_ttest) are a logical value indicating if p-value of post-test is smaller than global alpha.}
     \item{'strat_'}{All columns beginning with strat_ prefix are the evaluated decision strategies.}
}
}
\description{
This function simulates the impact of the assessement of statistical assumptions (pre-test) to decide on
a statistical hypothesis test (post-test).
}
\examples{
asm_simulate()
}
