% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_simulate.R
\name{asm_simVar}
\alias{asm_simVar}
\title{Simulate a single variable}
\usage{
asm_simVar(
  n = 30,
  distr = c("rnorm", "rnorm"),
  args = list(list(sd = 1, mean = 0), list(sd = 1, mean = 0)),
  share = c(0.3, 0.7),
  sorting = c("random", "grouped")[1]
)
}
\arguments{
\item{n}{integer samplesize to draw.}

\item{distr}{distribution function to draw sample with.}

\item{args}{args passed to distribution function.}

\item{share}{share of distributions.}

\item{sorting}{how to sort the simulated variable. 'grouped' means its sorted by the different distributions its sampled from.}
}
\description{
This function simulates contaminated variables as a mix of different distributions.
}
\examples{
asm_simVar()
}
