% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_ttest.R
\name{asm_ttest}
\alias{asm_ttest}
\title{t-Test with explicit assumptions}
\usage{
asm_ttest(x, y = NULL, ..., confirm = TRUE, diagnosePlots = TRUE, color = T)
}
\arguments{
\item{x}{numeric vector of data values.}

\item{y}{NULL or a numeric vector of data values.}

\item{...}{further arguments passed to t.test functions (like mu, paired).}

\item{confirm}{logical. If TRUE ask to confirm assumptions.}

\item{diagnosePlots}{logical. If TRUE print diagnostic plots.}

\item{color}{logical. If TRUE the assumption print is colored.}
}
\description{
This function is a wrapper for the  \link[stats]{t.test} function. Before conducting the t-Test diagnostic
plots are shown to assess the assumptions of the t-Test.
}
\examples{
asm_ttest(rnorm(30), rnorm(30, 2))
}
