% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_wilcox.R
\name{asm_wilcox}
\alias{asm_wilcox}
\title{Wilcoxon Test with explicit assumptions}
\usage{
asm_wilcox(x, y = NULL, ..., confirm = T, diagnosePlots = T, color = T)
}
\arguments{
\item{x}{numeric vector of data values.}

\item{y}{NULL or a numeric vector of data values.}

\item{...}{further arguments passed to t.test functions (like mu, paired).}

\item{confirm}{logical. If TRUE ask to confirm assumptions.}

\item{diagnosePlots}{logical. If TRUE print diagnostic plots.}

\item{color}{logical. If TRUE the assumption print is colored.}
}
\description{
This function is a wrapper for the \link[stats]{wilcox.test} function. Before conducting the wilcoxon test diagnostic
plots are shown to assess the assumptions of the wilcoxon test.
}
\examples{
asm_wilcox(rnorm(30), rnorm(30, 2))
}
