% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{assumptions}
\alias{assumptions}
\title{Print assumptions for commmon statistical tests}
\usage{
assumptions(call, color = T)
}
\arguments{
\item{call}{a function call for a hypothesis test like t.test.}

\item{color}{logical. If TRUE the assumption print is colored.}
}
\description{
This function prints assumptions of common statistical methods.
}
\examples{
assumptions(t.test)
assumptions(wilcox.test)
}
\references{
Conover, W. J. (1998). \emph{SOME METHODS BASED ON RANKS}. Chapter 5 of \emph{Practical nonparametric statistics}

Spanos, A. (2019). \emph{Misspecification (M-S) Testing}. Chapter 15 of \emph{Probability theory and statistical inference: Empirical modeling with observational data.}
}
