% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{check_as}
\alias{check_as}
\title{A Assumptions Checking Function}
\usage{
check_as(mod)
}
\arguments{
\item{mod}{Input a fitted linear model}
}
\description{
This functions check the assumptions of linear regression
}
\examples{
check_as(model), check_as(lm(y ~.))
check_as()
}
\keyword{linear,}
\keyword{regression}
