% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{summary_disp}
\alias{summary_disp}
\title{A Graphic Summary Function}
\usage{
summary_disp(modd)
}
\arguments{
\item{mod}{Input a fitted linear model}
}
\description{
This functions displays key graphs of linear regression
}
\examples{
summary_disp(model), summary_disp(lm(y ~ .))
summary_disp()
}
\keyword{graph}
\keyword{linear,}
\keyword{regression,}
